moved cache generator to tests as it will be used there and the reference with get.core.perf.js
